// src/components/baseButton/types.ts
export enum ButtonColor {
  Primary = 'primary',
  Secondary = 'secondary',
  Tertiary = 'tertiary',
  Error = 'error',
  White = 'white',
}

export enum ButtonVariant {
  Solid = 'solid',
  Outlined = 'outlined',
  Soft = 'soft',
  Ghost = 'ghost',
}

export enum ButtonShape {
  Rectangle = 'rectangle',
  Rounded = 'rounded',
  Tag = 'tag',
}

export enum ButtonTag {
  Button = 'button',
  Anchor = 'a',
}